% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getCharacter}
\alias{.getCharacter}
\title{.getCharacter}
\usage{
.getCharacter(x)
}
\arguments{
\item{x}{XML exported from Lattes imported to R as list.}
}
\value{
data frame.
}
\description{
Extract variable names in import process (get functions).
}
