% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtigosPublicados.R
\name{getArtigosPublicados}
\alias{getArtigosPublicados}
\title{getArtigosPublicados}
\usage{
getArtigosPublicados(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract published papers from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getArtigosPublicados(lattesXML[[499]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getArtigosPublicados)
 head(bind_rows(lt))
 }
}
}
