% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtuacoesProfissionais.R
\name{getAtuacoesProfissionais}
\alias{getAtuacoesProfissionais}
\title{getAtuacoesProfissionais}
\usage{
getAtuacoesProfissionais(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract profissional links from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getAtuacoesProfissionais(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getAtuacoesProfissionais)
 head(bind_rows(lt))
 }
}
