% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBancasDoutorado.R
\name{getBancasDoutorado}
\alias{getBancasDoutorado}
\title{getBancasDoutorado}
\usage{
getBancasDoutorado(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Ph.D. Examination Board's from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getBancasDoutorado(lattesXML[[499]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getBancasDoutorado)
 head(bind_rows(lt))
 }
}
}
