% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCapitulosLivros.R
\name{getCapitulosLivros}
\alias{getCapitulosLivros}
\title{getCapitulosLivros}
\usage{
getCapitulosLivros(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Books Chapter from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()) {
 data(xmlsLattes)
 # to import from one curriculum 
 getCapitulosLivros(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getCapitulosLivros)
 head(bind_rows(lt))
 }
}
