% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormacao.R
\name{getFormacao}
\alias{getFormacao}
\title{getFormacao}
\usage{
getFormacao(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
list
}
\description{
Extract Profissional Formation from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getFormacao(xmlsLattes[[499]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getFormacao)
 }
}
}
