% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormacaoMestrado.R
\name{getFormacaoMestrado}
\alias{getFormacaoMestrado}
\title{getFormacaoMestrado}
\usage{
getFormacaoMestrado(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
list
}
\description{
Extract Profissional Formation from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()) {
 data(xmlsLattes)
 # to import from one curriculum 
 getFormacao(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getFormacao)
 }
}
