% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getId.R
\name{getId}
\alias{getId}
\title{FUNCTION_TITLE}
\usage{
getId(curriculo)
}
\arguments{
\item{curriculo}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[xml2]{xml_attr}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}}
}
