% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdiomas.R
\name{getIdiomas}
\alias{getIdiomas}
\title{getIdiomas}
\usage{
getIdiomas(curriculo)
}
\arguments{
\item{curriculo}{'Lattes' XML imported as `xml2::read_xml()`.}
}
\value{
data frame
}
\description{
Extract Languages from 'Lattes' XML file.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()) {
 
 # to import from one curriculum 
 # curriculo <- xml2::read_xml('file.xml')
 # getIdiomas(curriculo)
 
 }
}
\seealso{
\code{\link[xml2]{xml_find_all}},\code{\link[xml2]{xml_children}},\code{\link[xml2]{xml_attr}}
 \code{\link[purrr]{map}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}}
 \code{\link[janitor]{clean_names}}
}
