% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrganizacaoEvento.R
\name{getOrganizacaoEvento}
\alias{getOrganizacaoEvento}
\title{getOrganizacaoEvento}
\usage{
getOrganizacaoEvento(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Event's Organization from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getOrganizacaoEvento(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getOrganizacaoEvento)
 head(bind_rows(lt))
 }
}
