% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrientacoesPosDoutorado.R
\name{getOrientacoesPosDoutorado}
\alias{getOrientacoesPosDoutorado}
\title{getOrientacoesPosDoutorado}
\usage{
getOrientacoesPosDoutorado(curriculo)
}
\description{
getOrientacoesPosDoutorado
}
