% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutrasProducoesTecnicas.R
\name{getOutrasProducoesTecnicas}
\alias{getOutrasProducoesTecnicas}
\title{getOutrasProducoesTecnicas}
\usage{
getOutrasProducoesTecnicas(curriculo)
}
\description{
getOutrasProducoesTecnicas
}
