% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParticipacaoProjeto.R
\name{getParticipacaoProjeto}
\alias{getParticipacaoProjeto}
\title{getParticipacaoProjeto}
\usage{
getParticipacaoProjeto(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Participation in Projects from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getParticipacaoProjeto(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getParticipacaoProjeto)
 head(bind_rows(lt))
 }
}
