% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrefacio.R
\name{getPrefacio}
\alias{getPrefacio}
\title{getPrefacio}
\usage{
getPrefacio(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Preface from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getPrefacio(lattesXML[[499]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getPrefacio)
 head(bind_rows(lt))
 }
}
}
