% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProgramaRadioTV.R
\name{getProgramaRadioTV}
\alias{getProgramaRadioTV}
\title{getProgramaRadioTV}
\usage{
getProgramaRadioTV(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract TV and Radio Program from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getProgramaRadioTV(xmlsLattes[[499]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getProgramaRadioTV)
 head(bind_rows(lt))
 }
}
}
