% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelatorioPesquisa.R
\name{getRelatorioPesquisa}
\alias{getRelatorioPesquisa}
\title{getRelatorioPesquisa}
\usage{
getRelatorioPesquisa(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Research Report from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getRelatorioPesquisa(xmlsLattes[[499]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getRelatorioPesquisa)
 head(bind_rows(lt))
 }
}
}
