% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrabalhosEventos.R
\name{getTrabalhosEventos}
\alias{getTrabalhosEventos}
\title{getTrabalhosEventos}
\usage{
getTrabalhosEventos(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Works in Event from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getTrabalhosEventos(lattesXML[[999]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getTrabalhosEventos)
 head(bind_rows(lt))
 }
}
}
