% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearizeNestedList.R
\name{LinearizeNestedList}
\alias{LinearizeNestedList}
\title{LinearizeNestedList}
\usage{
LinearizeNestedList(NList, LinearizeDataFrames = FALSE, NameSep = "/",
  ForceNames = FALSE)
}
\description{
LinearizeNestedList
}
