% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mostFrequent}
\alias{mostFrequent}
\title{mostFrequent}
\usage{
mostFrequent(x)
}
\arguments{
\item{x}{data frame.}
}
\value{
data frame.
}
\description{
Extract the most frequent ocorrence, helpfull to group_by.
}
