% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeYears.R
\name{normalizeYears}
\alias{normalizeYears}
\title{normalizeYears}
\usage{
normalizeYears(
  dataframe,
  year2normalize = "ano.do.artigo",
  issn = "issn",
  journalName = "titulo.do.periodico.ou.revista",
  paperTitle = "titulo.do.artigo"
)
}
\arguments{
\item{dataframe}{data imported with readLattes() then getArtigosPublicados()}

\item{year2normalize}{variable year}

\item{issn}{variable with issn, Default: 'issn'}

\item{journalName}{variable with journal name, Default: 'revista'}

\item{paperTitle}{variable with paper title, Default: 'titulo'}
}
\value{
data frame
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
if(interactive()){
data(latesXML)
l <- lapply(xmlsLattes, getArtigosPublicados) 
df <- bind_rows(al) \%>>\% tbl_df

ead(
    normalizeYears(dataframe=adf, 
              year2normalize='ano.do.artigo',
              issn='issn', 
              journalName= 'titulo.do.periodico.ou.revista',
              paperTitle='titulo.do.artigo') 
) 
}
}
\seealso{
\code{\link[dplyr]{group_by}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{mutate}}
}
