% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipLattes.R
\name{unzipLattes}
\alias{unzipLattes}
\title{unzipLattes}
\usage{
unzipLattes(filezip = "*.zip$", path = ".")
}
\arguments{
\item{filezip}{Zip files in your local machine.}

\item{path}{Directory with zip files.}
}
\description{
Unzip Lattes file to XML preserving 16 digits id as the file name.
}
\details{
Only zip files with 16 digits Lattes ids will be extracted.
}
\examples{
\dontrun{
if(interactive()){

unzipLattes(filezip='*.zip')
 
# to import from a diferent directory
# unzipLattes(filezip='*.zip', path='data/')
 }
}
}
