% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{xmlsLattes}
\alias{xmlsLattes}
\title{Two Lattes Curricula for example}
\format{
list
}
\usage{
xmlsLattes
}
\description{
Two Lattes curricula, from important researchers in my academic journey, imported as R list.
}
\keyword{Lattes}
