% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getApresentacaoTrabalho.R
\name{getApresentacaoTrabalho}
\alias{getApresentacaoTrabalho}
\title{getApresentacaoTrabalho}
\usage{
getApresentacaoTrabalho(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract academic's papers presentation, from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getApresentacaoTrabalho(lattesXML[[999]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getApresentacaoTrabalho)
 head(bind_rows(lt))
 }
}
}
