% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAreasAtuacao.R
\name{getAreasAtuacao}
\alias{getAreasAtuacao}
\title{getAreasAtuacao}
\usage{
getAreasAtuacao(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Research Area from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getAreasAtuacao(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getAreasAtuacao)
 head(bind_rows(lt))
 }
}
