% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBancasGraduacao.R
\name{getBancasGraduacao}
\alias{getBancasGraduacao}
\title{getBancasGraduacao}
\usage{
getBancasGraduacao(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Undergraduate Examination Board's from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()) {
 data(xmlsLattes)
 # to import from one curriculum 
 getBancasGraduacao(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getBancasGraduacao)
 head(bind_rows(lt))
 }
}
