% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBancasJulgadoras.R
\name{getBancasJulgadoras}
\alias{getBancasJulgadoras}
\title{getBancasJulgadoras}
\usage{
getBancasJulgadoras(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract General Examination Board's from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getBancasJulgadoras(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getBancasJulgadoras)
 head(bind_rows(lt))
 }
}
