% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBancasMestrado.R
\name{getBancasMestrado}
\alias{getBancasMestrado}
\title{getBancasMestrado}
\usage{
getBancasMestrado(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Master Examination Board's from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getBancasMestrado(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getBancasMestrado)
 head(bind_rows(lt))
 }
}
