% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnderecoProfissional.R
\name{getEnderecoProfissional}
\alias{getEnderecoProfissional}
\title{getEnderecoProfissional}
\usage{
getEnderecoProfissional(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Profissional Address from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getEnderecoProfissional(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getEnderecoProfissional)
 head(bind_rows(lt))
 }
}
