% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventosCongressos.R
\name{getEventosCongressos}
\alias{getEventosCongressos}
\title{getEventosCongressos}
\usage{
getEventosCongressos(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Events and Congresses from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getEventosCongressos(lattesXML[[999]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getEventosCongressos)
 head(bind_rows(lt))
 }
}
}
