% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormacaoGraduacao.R
\name{getFormacaoGraduacao}
\alias{getFormacaoGraduacao}
\title{getFormacaoGraduacao}
\usage{
getFormacaoGraduacao(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
list
}
\description{
Extract Profissional Formation from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()) {
 
 # to import from one curriculum 
 # curriculo <- xml2::read_xml('file.xml')
 # getFormacaoGraduacao(curriculo)  

 }
}
\seealso{
\code{\link[xml2]{xml_find_all}},\code{\link[xml2]{xml_attr}}
 \code{\link[purrr]{map}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}}
 \code{\link[janitor]{clean_names}}
}
