% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getId.R
\name{getId}
\alias{getId}
\title{getId}
\usage{
getId(curriculo)
}
\arguments{
\item{curriculo}{'Lattes' XML imported as `xml2::read_xml()`.}
}
\value{
character
}
\description{
Extract the unique 16 digits identification from 'Lattes' XML file.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()) {
 
 # to import from one curriculum 
 # curriculo <- xml2::read_xml('file.xml')
 # getId(curriculo)

 }
}
\seealso{
\code{\link[xml2]{xml_attr}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}}
}
