% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdiomas.R
\name{getIdiomas}
\alias{getIdiomas}
\title{getIdiomas}
\usage{
getIdiomas(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Languages from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getIdiomas(lattesXML[[499]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getIdiomas)
 head(bind_rows(lt))
 }
}
}
