% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJornaisRevistas.R
\name{getJornaisRevistas}
\alias{getJornaisRevistas}
\title{getJornaisRevistas}
\usage{
getJornaisRevistas(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Newspapers and Magazines from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getJornaisRevistas(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getJornaisRevistas)
 head(bind_rows(lt))
 }
}
