% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinhaPesquisa.R
\name{getLinhaPesquisa}
\alias{getLinhaPesquisa}
\title{getLinhaPesquisa}
\usage{
getLinhaPesquisa(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Research Lines from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getLinhaPesquisa(lattesXML[[499]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getLinhaPesquisa)
 head(bind_rows(lt))
 }
}
}
