% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLivrosPublicados.R
\name{getLivrosPublicados}
\alias{getLivrosPublicados}
\title{getLivrosPublicados}
\usage{
getLivrosPublicados(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Published Books from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getLivrosPublicados(lattesXML[[499]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getLivrosPublicados)
 head(bind_rows(lt))
 }
}
}
