% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrientacoesOutras.R
\name{getOrientacoesOutras}
\alias{getOrientacoesOutras}
\title{getOrientacoesOutras}
\usage{
getOrientacoesOutras(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Academic Advisory (Other) from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getOrientacoesOutras(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getOrientacoesOutras)
 head(bind_rows(lt))
 }
}
