% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrientacoesPosDoutorado.R
\name{getOrientacoesPosDoutorado}
\alias{getOrientacoesPosDoutorado}
\title{getOrientacoesPosDoutorado}
\usage{
getOrientacoesPosDoutorado(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Academic Advisory (Post-Doctorate) from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getOrientacoesPosDoutorado(lattesXML[[499]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getOrientacoesPosDoutorado)
 head(bind_rows(lt))
 }
}
}
