% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutrasProducoesBibliograficas.R
\name{getOutrasProducoesBibliograficas}
\alias{getOutrasProducoesBibliograficas}
\title{getOutrasProducoesBibliograficas}
\usage{
getOutrasProducoesBibliograficas(curriculo)
}
\description{
getOutrasProducoesBibliograficas
}
