% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParticipacaoProjeto.R
\name{getParticipacaoProjeto}
\alias{getParticipacaoProjeto}
\title{getParticipacaoProjeto}
\usage{
getParticipacaoProjeto(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Participation in Projects from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()) {
 
 # to import from one curriculum 
 # curriculo <- xml2::read_xml('file.xml')
 # getParticipacaoProjeto(curriculo)

 }
}
\seealso{
\code{\link[xml2]{xml_find_all}},\code{\link[xml2]{xml_attr}},\code{\link[xml2]{xml_children}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}}
 \code{\link[janitor]{clean_names}}
 \code{\link[tibble]{tibble}}
}
