% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPremiosTitulos.R
\name{getPremiosTitulos}
\alias{getPremiosTitulos}
\title{getPremiosTitulos}
\usage{
getPremiosTitulos(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Awards and Medals from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
if(interactive()){
 data(xmlsLattes)
 # to import from one curriculum 
 getPremiosTitulos(xmlsLattes[[2]])

 # to import from two or more curricula
 lt <- lapply(xmlsLattes, getPremiosTitulos)
 head(bind_rows(lt))
 }
}
