% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProducaoTecnica.R
\name{getProducaoTecnica}
\alias{getProducaoTecnica}
\title{getProducaoTecnica}
\usage{
getProducaoTecnica(curriculo)
}
\arguments{
\item{curriculo}{XML exported from Lattes imported to R as list.}
}
\value{
data frame
}
\description{
Extract Technical Production from XML file converted to R list.
}
\details{
Curriculum without this information will return NULL.
}
\examples{
\dontrun{
if(interactive()){
 data(lattesXML)
 # to import from one curriculum 
 getProducaoTecnica(lattesXML[[499]])

 # to import from two or more curricula
 lt <- lapply(lattesXML, getProducaoTecnica)
 head(bind_rows(lt))
 }
}
}
